Use new paypal checkout js to fix broken smart buttons
According to
https://developer.paypal.com/docs/checkout/reference/upgrade-integration
the old checkout.js script we were using has been deprecated, and we
need to use the new script.

The issue reported in #581 was an error coming from the deprecated
checkout.js script, and using the new script fixes that issue.

There are more uses of the deprecated script, but for now this seems to
address the most pressing issues.